requirements.txt
aiosmtplib==3.0.1
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.29.0
bcrypt==4.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
databases==0.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.1.2
fastapi==0.109.0
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.26.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
packaging==25.0
passlib==1.7.4
pillow==11.3.0
pluggy==1.6.0
polars==1.34.0
polars-runtime-32==1.34.0
psutil==7.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.6.1
pydantic_core==2.16.2
pyparsing==3.2.5
pytest==8.0.0
pytest-asyncio==0.23.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.2
requests==2.31.0
rsa==4.9.1
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.35.1
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
triton==3.4.0
typing_extensions==4.14.1
ultralytics==8.3.214
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.27.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
